Allow weights for a workout set to be higher than 999lbs
Believe it or not, I have a use-case for this! There's one particular exercise I do at my local gym -- the leg press -- where I do 1000lbs (sometimes slightly more) per rep.

I can sorta approximate this by just going with 999 (the current maximum allowable weight), but...it's an approximation.

This PR allows 4-digit numbers as weights for a workout set as long as that number begins with 1.